Fin de proyecto auth.js

+++ src/app/api/todos/route.ts
@@ -1,3 +1,4 @@
+import { getUserSessionServer } from "@/auth/actions/authActions";
 import prisma from "@/lib/prisma";
 import { NextResponse, NextRequest } from "next/server";
 import * as yup from "yup";
@@ -35,10 +36,17 @@ const postSchema = yup.object({
 });
 
 export async function POST(request: Request) {
+
+  const user = await getUserSessionServer()
+
+  if(!user) {
+    return NextResponse.json("No autorizado",{status:401});
+
+  }
   try {
     const {complete,description} = await postSchema.validate(await request.json());
 
-    const todo = await prisma.todo.create({ data: { complete,description} });
+    const todo = await prisma.todo.create({ data: { complete,description, userId: user.id} });
 
     return NextResponse.json(todo);
   } catch (error) {
@@ -48,9 +56,18 @@ export async function POST(request: Request) {
 
 
 export async function DELETE(request: Request) {
+
+  
+  const user = await getUserSessionServer()
+
+  if(!user) {
+    return NextResponse.json("No autorizado",{status:401});
+
+  }
+
   try {
 
-    const DeleteTodo = await prisma.todo.deleteMany({where:{complete:true}})
+    const DeleteTodo = await prisma.todo.deleteMany({where:{complete:true, userId:user.id}})
 
     return NextResponse.json("Borrados");
   } catch (error) {
